+++ aofl-js-packages/web-components/aofl-element/index.js
@@ -1,4 +1,4 @@
 export {AoflElement} from './modules/aofl-element';
 export * from './modules/decorators';
 export * from 'lit-element';
-export {render} from 'lit-html';
+export {render, templateCaches} from 'lit-html';