+++ cli-tools/cli/lib/webpack-config/__config/.babelrc.js
@@ -1,7 +1,22 @@
-const presets = ['@babel/preset-env'];
+const ignore = [
+  /[\/\\]core-js/,
+  /@babel[\/\\]runtime/,
+];
+
+const sourceType = 'unambiguous';
+
+const presets = [
+  [
+    "@babel/preset-env",
+    {
+      "modules": false,
+      "useBuiltIns": false
+    }
+  ]
+];
 const plugins = [
-  '@babel/plugin-proposal-object-rest-spread',
-  '@babel/plugin-syntax-dynamic-import',
+  // '@babel/plugin-proposal-object-rest-spread',
+  // '@babel/plugin-syntax-dynamic-import',
   '@babel/plugin-proposal-optional-chaining',
   [
     '@babel/plugin-proposal-decorators',
@@ -13,10 +28,11 @@ const plugins = [
   [
     '@babel/plugin-transform-runtime',
     {
-      'corejs': false,
-      'helpers': false,
-      'regenerator': true,
-      'useESModules': false
+      "absoluteRuntime": false,
+      "corejs": false,
+      "helpers": true,
+      "regenerator": true,
+      "useESModules": false
     }
   ]
 ];
@@ -36,6 +52,8 @@ if (process.env.NODE_ENV === 'test' && typeof process.env.SAUCE_USERNAME !== 'st
 }
 
 module.exports = {
+  ignore,
+  sourceType,
   presets,
   plugins
 };