+++ webpack-packages/templating-plugin/routes-config-loader/index.js
@@ -35,6 +35,61 @@ module.exports = async function() {
     })
       .replace(quoteRegex, '\'') + ';';
 
+    const hmrDeps = routeConfigObj.routes.map((item) => {
+      let str = item.resolve.replace(/\(\)\s*=>\s*import\('/gi, '\'');
+      str = str.replace(/'\)/g, '\'');
+      return str;
+    });
+
+    if (process.env.NODE_ENV !== 'production' && this.hot) {
+      content += `
+      const walk = function walk(root, call) {
+        call(root);
+        if (root.shadowRoot) {
+          walk(root.shadowRoot, call);
+        }
+
+        Array.from(root.children).forEach((child) => {
+          walk(child, call);
+        });
+      }
+
+      const hmr = function hmr() {
+        if (!module.hot) {
+          return;
+        }
+
+        module.hot.accept([${hmrDeps.join(', ')}], function (deps) {
+          const comp = __webpack_require__(deps[0]);
+          const Ctor = comp.default;
+
+          walk(document.body, (node) => {
+            if (node.localName === Ctor.is) {
+              const descriptorsS = Object.getOwnPropertyDescriptors(Ctor);
+              const descriptorsI = Object.getOwnPropertyDescriptors(Ctor.prototype);
+
+              for (const name in descriptorsS) {
+                if (name !== 'length' && name !== 'name' && name !== 'prototype') {
+                  Object.defineProperty(node.constructor, name, descriptorsS[name]);
+                }
+              }
+
+              for (const name in descriptorsI) {
+                Object.defineProperty(node, name, descriptorsI[name]);
+              }
+
+              if (node.connectedCallback) {
+                node.connectedCallback();
+                node.shouldUpdate(Ctor._classProperties);
+              }
+            }
+          });
+        });
+      }
+
+      hmr();
+      `;
+    }
     callback(null, content);
   } catch (e) {
     callback(e);