+++ CHANGELOG.md
@@ -10,9 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Much faster dev builds
 - @aofl/cli - serve added reporter option
 - @aofl/cli - made webpack mode available in .ejs context
-- @aofl/cli - webpack config supports sass
+- @aofl/cli - server supports hot and hot only options
 - @aofl/i18n - extracted core i18n functionality from the mixin
 - @aofl/web-components/aofl-element - added property decorator to support binding to store when declaring class properties
+- @aofl/hmr-loader
+- @aofl/web-components/aofl-element - store and state options to properties decorator
+- @aofl/cli - added cache-loader to webpack config
+- @aofl/templating-plugin - routes-config-loader supports hmr
 
 ### [Changed]
 - @aofl/cli - webpack config will not generate sourcemaps in dev by default
@@ -22,6 +26,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - @aofl/middleware - use() returns unsubscribe function
 - @aofl/router - hooks can unsubscribe
 - @aofl/webcomponent-css-loader - No longer imports global css into components and doesn't prune css in dev
+- @aofl/web-components - all components check customElements registry before calling define. This is done so hmr doesn't throw already defined error
+- @aofl/cli - updated babel config
+- @aofl/cli - a bunch of webpack optimization
 
 ### [Deprecated]
 ### [Removed]